backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
datasets>=2.0.0
transformers>=4.30.0
torch>=2.0.0
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
pytest>=7.4.0
httpx>=0.24.0